Initialise repository
As the project has grown, making it into a git repository has become a must. It is in a raw, totally not well thought state in this commit (yet it should build and run). It is hoped though that it will get better organised at some unspecified point in time.